Langchain_prompts/prompt_ui_static.py
from langchain_openai import ChatOpenAI
import streamlit as st
from dotenv import load_dotenv

load_dotenv()
model = ChatOpenAI()


st.header('Research Tool')

user_input = st.text_input('Enter Your Prompt')

if st.button('Summarize'):
    result = model.invoke(user_input)
    st.write(result.content)

